pyproject.toml
[project]
name = "graphrag-api"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "graphrag==2.2.1",
    "uvicorn",
    "pyyaml",
    "python-dotenv",
]
